maternity-risk-app/server/requirements.txt
# Flask==3.0.0
# gunicorn==21.2.0
# numpy==1.26.2
# pandas==2.1.4
# scikit-learn==1.3.2
# transformers==4.36.2
# torch
# huggingface-hub==0.20.3
# python-dotenv==1.0.0
# scipy==1.12.0
# joblib==1.3.2
# pickle-mixin==1.0.2
# requests==2.31.0
# werkzeug==3.0.1
# flask-cors==4.0.0
# setuptools>=68.0.0
# wheel>=0.40.0 

flask==3.0.0
gunicorn==21.2.0
requests==2.31.0
flask-cors==4.0.0
python-dotenv==1.0.0
